Error handling for frequency scaling

+++ include/types.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 namespace synergy {
-using frequency = unsigned long long;
+using frequency = unsigned;
 using power = unsigned long long;
 
 } // namespace synergy
